refactored homepage data
